refactor: move patches to utils for clarity
